use std::vector as return value for the FileReader's buffer readers.

+++ src/common/textures/animlib.h
@@ -105,7 +105,7 @@ struct anim_t
 //
 //****************************************************************************
 
-int32_t ANIM_LoadAnim(anim_t *anim, uint8_t *buffer, int32_t length);
+int32_t ANIM_LoadAnim(anim_t *anim, uint8_t *buffer, size_t length);
 
 //****************************************************************************
 //
